+++ feature/nameinput/src/main/kotlin/com/example/czechfoolapp/feature/nameinput/NameInputEvent.kt
@@ -2,5 +2,5 @@ package com.example.czechfoolapp.feature.nameinput
 
 sealed interface NameInputEvent {
     data class PlayerNameChanged(val id: Int, val value: String) : NameInputEvent
-    data class Next(val navigateToNext: () -> Unit) : NameInputEvent
+    data class Next(val navigateToNext: (Int) -> Unit) : NameInputEvent
 }
